Refactor common components folders
