+++ apps/payment-react-web/src/components/common/Balance/Balance.scss
