+++ apps/payment-react-web/src/components/common/Balance/index.ts
@@ -1,0 +1,1 @@
+export { Balance, BalancePure } from './Balance';
