+++ apps/payment-react-web/src/components/common/BlockchainLink/BlockchainLink.scss
@@ -1,4 +1,4 @@
-@use '../../styles/mixins.scss';
+@use '../../../styles/mixins.scss';
 
 $-compact-view-breakpoint: 500px;
