+++ apps/payment-react-web/src/components/common/BlockchainLink/index.ts
@@ -1,0 +1,1 @@
+export { BlockchainLink, BlockchainLinkPure, BlockchainLinkOption } from './BlockchainLink';
