+++ apps/payment-react-web/src/components/common/CopyButton/CopyButton.scss
@@ -1,6 +1,6 @@
 @use 'sass:math';
-@use '../../styles/layout-variables.scss';
-@use '../../styles/colors.scss';
+@use '../../../styles/layout-variables.scss';
+@use '../../../styles/colors.scss';
 
 .copy-button {
   margin: 0 math.div(layout-variables.$element-spacing-small, 2);
