+++ apps/payment-react-web/src/components/common/CopyButton/CopyButton.tsx
