+++ apps/payment-react-web/src/components/common/CopyButton/index.ts
@@ -1,0 +1,1 @@
+export { CopyButton, CopyButtonPure } from './CopyButton';
